rename purple_py -> purplepy
