+++ purplepy/vectorize.py
